+++ MiDAS/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 0.0.9043 - 22/04/20
++ fixes hlaToAAVariation failing on mixed resolution hla_calls input
++ adds warning when allele can not be found in the reference alignment
+
 ## 0.0.9042 - 17/04/20
 + removes unused pvalue_cutoff argument from runMiDAS function
 + fixes bug with conditional testing of glm models